+++ etc/blocks/monetization-test/block/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @milaboratories/milaboratories.monetization-test
 
+## 1.0.139
+
+### Patch Changes
+
+- @milaboratories/milaboratories.monetization-test.workflow@1.1.27
+
+## 1.0.138
+
+### Patch Changes
+
+- @milaboratories/milaboratories.monetization-test.workflow@1.1.26
+
+## 1.0.137
+
+### Patch Changes
+
+- @milaboratories/milaboratories.monetization-test.workflow@1.1.25
+
+## 1.0.136
+
+### Patch Changes
+
+- @milaboratories/milaboratories.monetization-test.workflow@1.1.24
+
+## 1.0.135
+
+### Patch Changes
+
+- @milaboratories/milaboratories.monetization-test.workflow@1.1.23
+
 ## 1.0.134
 
 ### Patch Changes