+++ etc/blocks/pool-explorer/block/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @milaboratories/milaboratories.pool-explorer
 
+## 1.0.319
+
+### Patch Changes
+
+- @milaboratories/milaboratories.pool-explorer.workflow@1.1.81
+
+## 1.0.318
+
+### Patch Changes
+
+- @milaboratories/milaboratories.pool-explorer.workflow@1.1.80
+
+## 1.0.317
+
+### Patch Changes
+
+- @milaboratories/milaboratories.pool-explorer.workflow@1.1.79
+
+## 1.0.316
+
+### Patch Changes
+
+- @milaboratories/milaboratories.pool-explorer.workflow@1.1.78
+
+## 1.0.315
+
+### Patch Changes
+
+- @milaboratories/milaboratories.pool-explorer.workflow@1.1.77
+
 ## 1.0.314
 
 ### Patch Changes