+++ etc/blocks/ui-examples/block/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @milaboratories/milaboratories.ui-examples
 
+## 1.0.408
+
+### Patch Changes
+
+- @milaboratories/milaboratories.ui-examples.workflow@1.1.81
+
+## 1.0.407
+
+### Patch Changes
+
+- @milaboratories/milaboratories.ui-examples.workflow@1.1.80
+
+## 1.0.406
+
+### Patch Changes
+
+- @milaboratories/milaboratories.ui-examples.workflow@1.1.79
+
+## 1.0.405
+
+### Patch Changes
+
+- @milaboratories/milaboratories.ui-examples.workflow@1.1.78
+
+## 1.0.404
+
+### Patch Changes
+
+- @milaboratories/milaboratories.ui-examples.workflow@1.1.77
+
 ## 1.0.403
 
 ### Patch Changes